requirements.txt
beautifulsoup4==4.12.3
debugpy==1.8.1
decorator==5.1.1
Django==4.2
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dvc==3.31.2
fqdn==1.5.1
google-generativeai
ipykernel==6.29.2
ipython==8.13.0
joblib==1.3.2
jupyterlab==4.1.2
keras==2.13.1
matplotlib==3.7.3
matplotlib-inline==0.1.6
notebook==7.1.2
numpy==1.23.3
pandas==2.0.3
pillow==10.2.0
psycopg2-binary==2.9.9
pydot==2.0.0
PyJWT==1.7.1
pyparsing==3.1.1
python-box==6.0.2
python-dotenv==1.0.1
# pytz==2022.6
# pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.3.1
scipy==1.10.1
sqlparse==0.4.3
sqltrie==0.11.0
tensorflow==2.13.0
tqdm==4.66.1
types-PyYAML==6.0.12.12
typing_extensions==4.5.0
urllib3==2.2.1
dj-database-url==2.1.0
gunicorn==21.2.0
